--- a/Assets/Scripts/Internal/DaggerfallBillboardSystem.cs.meta
+++ b/Assets/Scripts/Internal/DaggerfallBillboardSystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2b85dcd99c159ca419ee965c962e7bff
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: