Hi guys,

This is my first attempt here at implementing something that #2316 enabled.
I picked billboards not because those were causing any mayor performance issues, **they weren't**, but because I knew how to optimize them already and it is pretty much an isolated system from many complexities of this project.

<br>

These are **very** rough (editor & debug) numbers how billboards performed
> **Before**: `0.6 ms` (main thread)

![Screenshot 2022-07-12 181808](https://user-images.githubusercontent.com/3066539/179237191-bc10e0d3-1f9f-4ad5-9d06-3fb52f8f01f8.png)

> **After**:  `0.03 ms` (main thread) + `0.06 ms` (different thread)

![Screenshot 2022-07-13 120343](https://user-images.githubusercontent.com/3066539/179237192-31d74c57-3529-4e3f-bb5f-f4613d93fcd5.png)

![Screenshot 2022-07-13 120516](https://user-images.githubusercontent.com/3066539/179237189-64fab5bc-fe4c-4878-8090-d1a5f61562a2.png)
> ☝ different thread

We can see that raw computation cost decreased about 10x, but that is not even that important. The thing that matters here is that this `0.5 ms` disappeared from the main thread, so it will no longer affect the main gameplay/rendering loop from now on.

<br>

If you point me any cases which I should test and make sure it works, that would be great. It looks fine on my machine, but I want to make sure my changes won't break something that I didn't even consider.