@Interkarma 2022-07-15
Thank you for PR. It's interesting to see this as a test example, though I'm not convinced a 0.5ms saving per frame is worth the additional complexity. The PR also changes code for no reason, such as removing pattern to restart anims after disable and introducing alternate alignment methods. I'd prefer to see a much lighter touch here focused on optimisation effort only.

Please also keep in mind that I'm trying to reach 1.0 milestone, so any refactors that don't serve bug fix or localisation aren't on my radar right now. I'm still open to small changes, particularly for mod support, but they have to be very light touch.

It would be interesting to compare the savings in axial alignment using a shader (similar to what I did with billboard batch). Though I'm not sure it would be worth the additional complexity at this point either.