@Interkarma 2022-07-16
Thank you for refinement. The animation stall is now resolved and easier overall for me to review the optimisation changes.

This does make a quite a difference to the billboard alignment times. I profiled across locations of all sizes and found the 0.5ms per frame to be a very consistent saving. I can see how this setup could be useful in other places.

I'm a bit hesitant to merge right now. I'm committed to a stable 0.14.3 release by next week and don't want to introduce any last-second changes. But the process appears good and now doesn't introduce any problems I'm able to find. I'll work on merging this one sometime in next few releases.