@andrew-raphael-lukasik 2022-09-05
I have doubts about this PR - is it really a good enough of a solution. Mainly because that occlusion-related overhead since [a8a36c4](https://github.com/Interkarma/daggerfall-unity/pull/2389/commits/a8a36c44f5b61423dc50e18460d4dbbeff243db5).